Replace hilt with Koin (which is multi-platform dependency injection)

+++ frontend/src/androidMain/kotlin/org/robphi/groupnotes/model/Models.kt
@@ -1,6 +1,6 @@
 package org.robphi.groupnotes.model
 
 data class Note(
-    val noteId: Int,
+    val noteId: Long,
     val title: String
 )
